extracting and cleaning
